refactor: split transpiler into multiple files
